refactor: consolidate updateOrderInStorage to LSPStore
